tests/wpt/web-platform-tests/workers/interfaces/WorkerUtils/importScripts/012.html

<!doctype html>
<title>importScripts(1)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('012.js');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, '1');
    this.done();
  });
  worker.onerror = this.step_func(function(e) {
    assert_unreached(e.message);
  });
});
</script>
